Added click cooldowns that are visualized

+++ src/main/java/engine/ecs/serialization/dataStructures/ResourceData.java
@@ -1,4 +1,4 @@
 package engine.ecs.serialization.dataStructures;
 
-public record ResourceData(int uid, String name, int amount, String type) {
+public record ResourceData(int uid, String name, float amount, String type) {
 }
